perf: replace DOUBLE with FLOAT to eliminate double-precision soft-float (~2KB flash saved)
- cli_float.h/cli_float.c: double -> float
- cli_parse.h/cli_parse.c: cli_parse_double -> cli_parse_float
- cli_var.c: builtin_double_* -> builtin_float_*
- cmd_dispose.h/cmd_dispose.c: CLI_TYPE_DOUBLE -> CLI_TYPE_FLOAT
- cli_errno_dispose.h/cli_errno.c: CLI_ERR_DOUBLE_* -> CLI_ERR_FLOAT_*
- tests: update test_double.c and test_cli_var.c accordingly

Cortex-M4 FPU now handles float ops in hardware (vadd.f32/vmul.f32),
eliminating all __aeabi_dmul/dadd/dsub/dcmp libgcc dependencies.

+++ tests/test_cli_var.c
@@ -22,8 +22,8 @@ CLI_VAR(g_device_name, "g_device_name", STRING, "Device name");
 static bool g_verbose = false;
 CLI_VAR(g_verbose, "g_verbose", BOOL, "Verbose output flag");
 
-static double g_kp = 2.5;
-CLI_VAR_RO(g_kp, "g_kp", DOUBLE, "PID Kp parameter (read-only)");
+static float g_kp = 2.5f;
+CLI_VAR_RO(g_kp, "g_kp", FLOAT, "PID Kp parameter (read-only)");
 
 /* ============================================================
  * 自定义类型测试：二维坐标点
